refactor: Add routers for datasets and add endpoints
Added routers for datasets and add endpoints

Refactor #COG-334
